Abstract creator test classes to share common logic between creators
